finish redirections, push_tokens, and organize folder
