+++ Makefile
@@ -6,7 +6,7 @@
 #    By: migmanu <marvin@42.fr>                     +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2023/10/07 19:24:44 by migmanu           #+#    #+#              #
-#    Updated: 2023/11/09 18:35:25 by johnavar         ###   ########.fr        #
+#    Updated: 2023/11/15 18:27:55 by sebasnadu        ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
@@ -56,7 +56,7 @@ $(NAME): $(OBJ_FILES) $(LIBFT_PATH)
 	@echo $(GREEN) "OK COMPILED" $(EOC)
 
 $(LIBFT_PATH):
-	@$(MAKE) -C $(LIBFT_DIR) -s
+	@make bonus -C $(LIBFT_DIR) -s
 
 %.o: %.c
 	@git submodule update --init -q
