+++ include/minishell.h
@@ -6,7 +6,7 @@
 /*   By: jmigoya- <jmigoya-@student.42berlin.de>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/04 14:22:24 by migmanu           #+#    #+#             */
-/*   Updated: 2023/11/14 18:34:29 by sebasnadu        ###   ########.fr       */
+/*   Updated: 2023/11/15 18:13:16 by sebasnadu        ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -24,6 +24,7 @@
 # include <sys/types.h>
 # include <sys/wait.h>
 # include <errno.h>
+# include <sys/ioctl.h>
 
 // I remplaced them for the enum below, to iteract with the err msg.
 // we leave here if we want to come back.
@@ -31,17 +32,20 @@
 /*# define FAILURE 1*/
 /*# define SUCCESS 0*/
 
+# define WRITE_END 1
+# define READ_END 0
 // in Test
 int	g_exit_status;
 
 enum e_mish_err
 {
-	ERROR = -1,
 	SUCCESS,
 	FAILURE,
+	ERROR,
 	INV_ARGS,
 	FORK_ERR,
-	UNQUOTE
+	PIPE_ERR,
+	UNQUOTE,
 	NO_PERM = 126,
 	NO_FILE = 127,
 };
@@ -100,15 +104,33 @@ char				**insert_subwords(char ***mtx, char **new_items, int pos);
 // expander_utils
 char				*expand_home(char *str, int i, int quotes[2], char *home);
 char				*expand_vars(t_data *mish, char *str, int quotes[2], int i);
-// utils
+// utils.c
 int					count_subwords(char *s, char *set, int count);
 int					count_words(const char *str, char *set, int cts[2]);
 int					ft_matrixlen(char **matrix);
 int					find_inset(const char *str, char *set);
+char				**ft_matrixdup(char **tokens);
+// syntax_list_utils.c
+char				**trim_all(char **tokens);
+char				*get_trimmed_str(char *str);
+int					len_noquotes(char *str);
+t_scmd				*init_node(void);
+t_list				*clean_fail(t_list *cmds, char **tokens, char **tmp);
+// get_node.c
+t_scmd				*get_node(t_scmd *node, char **cmds[2], int *i);
+// redirections.c
+t_scmd				*redir_in(t_scmd *node, char **cmds, int *i);
+t_scmd				*redir_in_heredoc(t_scmd *node, char **cmds, int *i);
+t_scmd				*redir_out(t_scmd *node, char **cmds, int *i);
+t_scmd				*redir_out_append(t_scmd *node, char **cmds, int *i);
+// redirections_utils.c
+int					get_heredoc_fd(char *limit);
+int					get_fd(int oldfd, char *path, int flags[2]);
 
 // exit.c
 void				mish_error(t_data *mish, char *param, int err, int is_exit);
-void				ft_free_matrix(char ***matrix);
+void				ft_matrixfree(char ***matrix);
+void				free_scmd(void *content);
 
 // init
 char				*init_prompt(t_data *mish);
