+++ src/main.c
@@ -6,7 +6,7 @@
 /*   By: migmanu <marvin@42.fr>                     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/04 15:50:52 by migmanu           #+#    #+#             */
-/*   Updated: 2023/11/12 02:23:06 by sebasnadu        ###   ########.fr       */
+/*   Updated: 2023/11/15 16:15:27 by sebasnadu        ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -28,6 +28,11 @@ int	main(int argc, char *argv[], char *envp[])
 		line = init_prompt(&mish);
 		input_handler(line, &mish);
 		//executor_router(&mish);
+		// TODO: this clean all the cmds and also close fds of files but, i think
+		// it should be provisional for now, this should be after the executor but inside
+		// your function to have more cleaning main
+		if (mish.cmds)
+			ft_lstclear(&mish.cmds, free_scmd);
 	}
 	return (0);
 }
