+++ src/tests/exit.c
@@ -6,13 +6,13 @@
 /*   By: johnavar <johnavar@student.42berlin.d      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/09 16:45:19 by johnavar          #+#    #+#             */
-/*   Updated: 2023/11/14 19:42:58 by sebasnadu        ###   ########.fr       */
+/*   Updated: 2023/11/15 13:37:47 by sebasnadu        ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../../include/minishell.h"
 
-void	ft_free_matrix(char ***matrix)
+void	ft_matrixfree(char ***matrix)
 {
 	int	i;
 
@@ -29,15 +29,29 @@ void	ft_free_matrix(char ***matrix)
 	}
 }
 
+void	free_scmd(void *content)
+{
+	t_scmd	*node;
+
+	node = content;
+	ft_matrixfree(&node->full_cmd);
+	free(node->full_cmd);
+	if (node->in_fd != STDIN_FILENO)
+		close(node->in_fd);
+	if (node->out_fd != STDOUT_FILENO)
+		close(node->out_fd);
+	free(node);
+}
+
 // TODO: the exit code its wrong, maybe we can implement a global with
 // the las exit and use here to exit correct.
 void	mish_error(t_data *mish, char *param, int err, int is_exit)
 {
-	if (err == PIPE_ERR || err == FORK_ERR)
+	if (err == FORK_ERR || err == PIPE_ERR)
 		g_exit_status = 1;
 	else
 		g_exit_status = err;
-	if (err != SUCCESS && err != FAILURE)
+	if (err != SUCCESS && err != FAILURE && err != ERROR)
 	{
 		ft_putstr_fd("mish: ", STDERR_FILENO);
 		if (err == INV_ARGS)
@@ -57,8 +71,12 @@ void	mish_error(t_data *mish, char *param, int err, int is_exit)
 		ft_putstr_fd("\n", STDERR_FILENO);
 	}
 	if (mish && is_exit)
+	{
 		if (mish->env)
 			hashmap_free_table(mish->env);
-	else if (is_exit)
+		if (mish->cmds)
+			ft_lstclear(&mish->cmds, free_scmd);
+	}
+	if (is_exit)
 		exit(g_exit_status);
 }
