+++ src/tests/signals.c
@@ -6,7 +6,7 @@
 /*   By: migmanu <marvin@42.fr>                     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/06 15:22:11 by migmanu           #+#    #+#             */
-/*   Updated: 2023/11/06 18:43:24 by migmanu          ###   ########.fr       */
+/*   Updated: 2023/11/15 17:30:08 by sebasnadu        ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,13 +14,22 @@
 #include <readline/readline.h>
 #include <signal.h>
 
+// INFO: testing ioctl with TIOCSTI for simulate a new line from user
+// a normal \n doesnt work good with here_doc,
+// rl_on_new_line and not rl_redisplay, because its not necessary redisplays all
+// and it cause a extra space on here_doc
 void	handle_sigint(int signum)
 {
-	ft_putchar_fd('\n', 1);
-	rl_on_new_line();
-	rl_replace_line("", 0);
-	rl_redisplay();
-	(void)signum;
+	if (signum == SIGINT)
+	{
+		g_exit_status = 130;
+		ioctl(STDIN_FILENO, TIOCSTI, "\n");
+		// ft_putchar_fd('\n', 1);
+		rl_replace_line("", 0);
+		rl_on_new_line();
+		// rl_redisplay();
+	}
+	// (void)signum;
 }
 
 void	config_signals(void)
